Refactor: 메뉴 enum 클래스 menu패키지로 이동
