+++ src/main/java/com/sprint/mission/discodeit/JavaApplication.java
@@ -3,6 +3,10 @@ package com.sprint.mission.discodeit;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.User;
+import com.sprint.mission.discodeit.menu.ChannelMenu;
+import com.sprint.mission.discodeit.menu.MainMenu;
+import com.sprint.mission.discodeit.menu.MessageMenu;
+import com.sprint.mission.discodeit.menu.UserMenu;
 import com.sprint.mission.discodeit.repository.file.FileChannelRepository;
 import com.sprint.mission.discodeit.repository.file.FileMessageRepository;
 import com.sprint.mission.discodeit.repository.file.FileUserRepository;
